requirements.txt
streamlit == 1.19.0
folium == 0.14.0
streamlit_folium == 0.11.1
pandas == 1.4.4
pillow == 9.2.0
inflection == 0.5.1
plotly == 5.9.0